requirements.txt
Flask>=2.0
openpyxl>=3.0
psycopg2-binary
gunicorn